extract checkpoint helpers
